test: make websocket expiry regression deterministic
