+++ tests/websocket_lifecycle.rs
@@ -1543,141 +1543,6 @@ async fn inactive_subscription_expires_after_five_minutes_despite_heartbeats() {
 }
 
 #[tokio::test]
-async fn partial_update_reports_applied_symbols_at_absolute_expiry() {
-    let listener = TcpListener::bind("127.0.0.1:0")
-        .await
-        .expect("bind mock WebSocket server");
-    let endpoint = format!("ws://{}", listener.local_addr().expect("listener address"));
-    let (add_seen, add_received) = tokio::sync::oneshot::channel();
-    let server = tokio::spawn(async move {
-        let (stream, _) = listener.accept().await.expect("accept client");
-        let mut socket = accept_async(stream).await.expect("WebSocket handshake");
-        socket
-            .next()
-            .await
-            .expect("subscribe command")
-            .expect("valid subscribe command");
-        socket
-            .send(Message::text(
-                json!({
-                    "messageType": "I",
-                    "data": {"subscriptionId": 47},
-                    "response": {"code": 200, "message": "subscribed"}
-                })
-                .to_string(),
-            ))
-            .await
-            .expect("send acknowledgement");
-        let mut heartbeat = tokio::time::interval(Duration::from_secs(30));
-        let mut removed = false;
-        let mut add_seen = Some(add_seen);
-        loop {
-            tokio::select! {
-                _ = heartbeat.tick() => {
-                    if socket.send(Message::text(json!({
-                        "messageType": "H",
-                        "response": {"code": 200, "message": "heartbeat"}
-                    }).to_string())).await.is_err() {
-                        return;
-                    }
-                }
-                incoming = socket.next() => match incoming {
-                    Some(Ok(Message::Text(payload))) if !removed => {
-                        let payload: serde_json::Value =
-                            serde_json::from_str(&payload).expect("remove JSON");
-                        assert_eq!(payload["eventName"], "unsubscribe");
-                        socket.send(Message::text(json!({
-                            "messageType": "I",
-                            "data": {"subscriptionId": "expiry-remove-id"},
-                            "response": {"code": 200, "message": "updated"}
-                        }).to_string())).await.expect("acknowledge remove");
-                        removed = true;
-                    }
-                    Some(Ok(Message::Text(payload))) => {
-                        let payload: serde_json::Value =
-                            serde_json::from_str(&payload).expect("add JSON");
-                        if let Some(add_seen) = add_seen.take() {
-                            assert_eq!(payload["eventName"], "subscribe");
-                            add_seen.send(()).expect("record pending add");
-                        }
-                    }
-                    Some(Ok(Message::Close(_))) | None => return,
-                    Some(Ok(_)) => {}
-                    Some(Err(_)) if add_seen.is_none() => return,
-                    Some(Err(error)) => panic!("receive client message: {error}"),
-                }
-            }
-        }
-    });
-    let connector = TiingoConnector::with_endpoints(endpoint.clone(), endpoint)
-        .expect("mock endpoint is valid");
-    let registry = MarketDataRegistry::with_connector(Some("test-key".into()), connector);
-    let started = registry
-        .start(StartRequest {
-            service: Service::Iex,
-            symbols: vec!["AAPL".into(), "MSFT".into()],
-            threshold_level: None,
-            confirm_iex_market_data_agreement: false,
-        })
-        .await
-        .expect("subscription starts");
-
-    tokio::time::pause();
-    for _ in 0..29 {
-        tokio::time::advance(Duration::from_secs(60)).await;
-        for _ in 0..10 {
-            tokio::task::yield_now().await;
-        }
-        registry
-            .poll_with_bounds(&started.id, 0, 1, Duration::ZERO)
-            .await
-            .expect("minute touch succeeds");
-    }
-    assert!(
-        !server.is_finished(),
-        "activity keeps the idle deadline refreshed"
-    );
-    tokio::time::advance(Duration::from_secs(56)).await;
-    let update_registry = registry.clone();
-    let update_id = started.id.clone();
-    let update = tokio::spawn(async move {
-        update_registry
-            .update(
-                &update_id,
-                UpdateRequest {
-                    add_symbols: vec!["NVDA".into()],
-                    remove_symbols: vec!["AAPL".into()],
-                    threshold_level: None,
-                },
-            )
-            .await
-    });
-    tokio::time::resume();
-    add_received.await.expect("server sees pending add");
-    tokio::time::pause();
-    tokio::time::advance(Duration::from_secs(4)).await;
-    let error = update
-        .await
-        .expect("update task joins")
-        .expect_err("absolute expiry interrupts the pending add");
-    let error = serde_json::to_value(error.payload()).expect("serialize expiry update error");
-    assert_eq!(error["kind"], "timeout");
-    assert_eq!(error["appliedSymbols"], json!(["MSFT"]));
-    tokio::time::resume();
-    tokio::time::timeout(Duration::from_secs(1), server)
-        .await
-        .expect("absolute expiry closes the socket promptly")
-        .expect("mock server exits cleanly");
-    let expired = registry
-        .poll_with_bounds(&started.id, 0, 1, Duration::ZERO)
-        .await
-        .expect("expiry remains inspectable");
-    assert_eq!(expired.state, SubscriptionStatus::Expired);
-
-    registry.shutdown().await;
-}
-
-#[tokio::test]
 async fn liveness_fails_after_seventy_five_seconds_and_heartbeat_and_data_refresh_it() {
     let listener = TcpListener::bind("127.0.0.1:0")
         .await
